pyproject.toml
[project]
name = "claim-processing-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.12",
    "langchain-community>=0.3.21",
    "langchain-core==0.3.51",
    "langchain-openai==0.3.12",
    "langchain-siliconflow>=0.1.1",
    "langgraph==0.4.1",
    "pydantic==2.11.1",
    "pydantic-settings==2.9.1",
    "pymupdf==1.25.5",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "uvicorn[standard]==0.34.2",
]
